cleanup(socha): code of pollution and generalized it
